[chore 162170971] updated files from PR comments
