+++ src/index.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './routes/App.jsx';
+import App from './App.jsx';
 
 
 ReactDOM.render(
